tidy up some getApiContestData loose ends
This came out of some troubleshooting attempts lately when the fetch
doesn't always work. Specific things addressed:

1. usage from gatsby-node.esm.js breaks if Typescript syntax is actually used, so I changed the file extension to .js
2. but I did still annotate the types, using various JSDoc syntax that VSCode supports (but doesn't really enforce)
3. all the current callers crash-and-burn if the original catch block is triggered and returns undefined; returning an empty array in that case still tends to lead to trouble down the road and seems wrong. so now this bubbles the original error up instead of waiting for some sort of method-call-on-undefined error later
4. also no longer tries to return the parsed JSON if the API server returned some sort of unexpected status code, but rather logs the response body as text to aid troubleshooting
